+++ tests/test_determinism.py
@@ -133,7 +133,7 @@ def _test_determinism_matsim(index, data_path, tmpdir):
         #"ile_de_france_network.xml.gz":     "5f10ec295b49d2bb768451c812955794",
         "ile_de_france_households.xml.gz":  "64a0c9fab72aad51bc6adb926a1c9d44",
         #"ile_de_france_facilities.xml.gz":  "5ad41afff9ae5c470082510b943e6778",
-        "ile_de_france_config.xml":         "f374807f12a5151fe1efb6e9904e1a56"
+        "ile_de_france_config.xml":         "481fac5fb3e7b90810caa38ff460c00a"
     }
 
     # activities.gpkg, trips.gpkg, meta.json,